adicionando grafico de barras em overview
